Extract Field and Part handlers
