+++ AndroidAnnotations/androidannotations-rest-spring/rest-spring/src/main/java/org/androidannotations/rest/spring/RestSpringPlugin.java
@@ -22,11 +22,13 @@ import org.androidannotations.AndroidAnnotationsEnvironment;
 import org.androidannotations.handler.AnnotationHandler;
 import org.androidannotations.plugin.AndroidAnnotationsPlugin;
 import org.androidannotations.rest.spring.handler.DeleteHandler;
+import org.androidannotations.rest.spring.handler.FieldHandler;
 import org.androidannotations.rest.spring.handler.GetHandler;
 import org.androidannotations.rest.spring.handler.HeadHandler;
 import org.androidannotations.rest.spring.handler.HeaderHandler;
 import org.androidannotations.rest.spring.handler.HeadersHandler;
 import org.androidannotations.rest.spring.handler.OptionsHandler;
+import org.androidannotations.rest.spring.handler.PartHandler;
 import org.androidannotations.rest.spring.handler.PathHandler;
 import org.androidannotations.rest.spring.handler.PostHandler;
 import org.androidannotations.rest.spring.handler.PutHandler;
@@ -46,6 +48,8 @@ public class RestSpringPlugin extends AndroidAnnotationsPlugin {
 	public List<AnnotationHandler<?>> getHandlers(AndroidAnnotationsEnvironment androidAnnotationEnv) {
 		List<AnnotationHandler<?>> annotationHandlers = new ArrayList<>();
 		annotationHandlers.add(new RestHandler(androidAnnotationEnv));
+		annotationHandlers.add(new FieldHandler(androidAnnotationEnv));
+		annotationHandlers.add(new PartHandler(androidAnnotationEnv));
 		annotationHandlers.add(new GetHandler(androidAnnotationEnv));
 		annotationHandlers.add(new PostHandler(androidAnnotationEnv));
 		annotationHandlers.add(new PutHandler(androidAnnotationEnv));
